# api/swagger.yaml
...
swagger: '2.0'
info:
  description: 'SRS and workflow: https://docs.google.com/document/d/17aiIiosJApvnAliPajrcKnusvc9YdgLgiccv_TbMRGM

'
  version: 1.0.0
  title: Monero Wallet service API
basePath: "/"
schemes:
- http
- https
paths:
...
  "/api/v1/info/wallet":
    get:
      summary: Wallet Info
      description: Get wallet status
      tags:
      - Info
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      responses:
        '200':
          description: Status 200
          schema:
            required:
            - status
            - participants
            - joined
            - changed_keys
            type: object
            properties:
              status:
                type: string
              participants:
                type: integer
              joined:
                type: integer
              changed_keys:
                type: integer
  "/api/v1/info/multisig":
    get:
      summary: Multisig Info
      tags:
      - Info
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string

      responses:
        '200':
          description: Status 200
          schema:
            required:
            - multisig_infos
            type: object
            properties:
              multisig_infos:
                type: array
                items:
                  required:
                  - multisig_info
                  type: object
                  properties:
                    multisig_info:
                      type: string
  "/api/v1/info/extra_multisig":
    get:
      summary: Push Extra Multisig Info
      description: Pushes via websocket
      tags:
      - Info
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      responses:
        '200':
          description: Status 200
          schema:
            required:
            - extra_multisig_infos
            type: object
            properties:
              extra_multisig_infos:
                type: array
                items:
                  required:
                  - extra_multisig_info
                  type: object
                  properties:
                    extra_multisig_info:
                      type: string
  "/api/v1/tx_proposals":
    post:
      summary: create transaction proposal
      tags:
      - Tx Proposal
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - amount
          - destination_address
          - fee
          - proposal_id
          - signed_transaction
          properties:
            destination_address:
              type: string
              description: valid monero address
            description:
              type: string
              description: human readable descritption
            signed_transaction:
              type: string
              description: serialized transaction signed by initiator
            amount:
              type: integer
              description: transaction amount in atomic monero units
            fee:
              type: integer
              description: transaction fee in atomic monero units
      responses:
        '200':
          description: Proposal created
          schema:
            type: object
            properties:
              proposal_id:
                type: string
                description: id of created proposal
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
    get:
      summary: get current proposals
      tags:
      - Tx Proposal
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: signing
        in: query
        required: false
        type: boolean
      responses:
        '200':
          description: Status 200
          schema:
            type: array
            items:
              required:
              - amount
              - approvals
              - destination_address
              - fee
              - proposal_id
              - rejects
              - status
              type: object
              properties:
                proposal_id:
                  type: integer
                destination_address:
                  type: string
                  description: valid monero address
                amount:
                  type: integer
                  description: transaction amount in atomic monero units
                fee:
                  type: integer
                  description: transaction fee in atomic monero units
                approvals:
                  type: array
                  items:
                    type: string
                rejects:
                  type: array
                  items:
                    type: string
                description:
                  type: string
                  example: donation to poor c++ developer
                status:
                  type: string
                  enum:
                  - signing
                  - success
                  - failed
                  - rejected
        '403':
          description: Status 403
  "/api/v1/tx_proposals/{proposal_id}":
    get:
      summary: get transaction proposal
      tags:
      - Tx Proposal
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: proposal_id
        in: path
        required: true
        type: string
      responses:
        '200':
          description: Status 200
          schema:
            required:
            - amount
            - approvals
            - destination_address
            - fee
            - proposal_id
            - rejects
            - status
            type: object
            properties:
              proposal_id:
                type: integer
              destination_address:
                type: string
                description: valid monero address
              amount:
                type: integer
                description: transaction amount in atomic monero units
              fee:
                type: integer
                description: transaction fee in atomic monero units
              approvals:
                type: array
                items:
                  type: string
              rejects:
                type: array
                items:
                  type: string
              description:
                type: string
                example: donation to poor c++ developer
              status:
                type: string
                enum:
                - signing
                - success
                - failed
                - rejected
        '403':
          description: Status 403
        '404':
          description: Status 404
  "/api/v1/tx_proposals/{proposal_id}/decision":
    head:
      summary: check proposal decision lock
      tags:
      - Tx Proposal
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: proposal_id
        in: path
        required: true
        type: string
      responses:
        '200':
          schema:
            required:
            - status
            type: object
            properties:
              status:
                type: string
                enum:
                - ok
        '409':
          description: Status 409, Lock can't be acquired
        '404':
          description: Status 404
    put:
      summary: sign of dismiss transaction
      tags:
      - Tx Proposal
      consumes:
      - application/json
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: proposal_id
        in: path
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - approved
          properties:
            signed_transaction:
              type: string
              description: serialized transaction and signed transaction
            approved:
              type: boolean
            approval_nonce:
              type: integer
              description: amount of approvals for proposal to check if signing actual transaction
      responses:
        '204':
          description: Decision accepted
        '400':
          description: Status 400
          schema:
            required:
            - error
            type: object
            properties:
              error:
                type: string
        '403':
          description: Status 403
        '404':
          description: Status 404
  "/api/v1/tx_relay_status/{proposal_id}":
    post:
      summary: set proposal transaction id after it was relayed
      tags:
      - Tx Proposal
      produces:
      - application/json
      parameters:
      - name: X-Session-Id
        in: header
        required: true
        type: string
      - name: X-Nonce
        in: header
        required: true
        type: integer
        minimum: 0
      - name: X-Signature
        in: header
        required: true
        type: string
      - name: proposal_id
        in: path
        required: true
        type: string
      - name: body
        in: body
        required: true
        schema:
          type: object
          required:
          - tx_id
          properties:
            tx_id:
              type: array
              items:
                type: string
              description: relayed transaction id (list)
      responses:
        '200':
          description: Status 200
          schema:
            required:
            - amount
            - approvals
            - destination_address
            - fee
            - proposal_id
            - rejects
            - status
            type: object
            properties:
              proposal_id:
                type: integer
              destination_address:
                type: string
                description: valid monero address
              amount:
                type: integer
                description: transaction amount in atomic monero units
              fee:
                type: integer
                description: transaction fee in atomic monero units
              approvals:
                type: array
                items:
                  type: string
              rejects:
                type: array
                items:
                  type: string
              description:
                type: string
                example: donation to poor c++ developer
              status:
                type: string
                enum:
                - signing
                - success
                - failed
                - rejected
        '403':
          description: Status 403
        '404':
          description: Status 404
...